apps/api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
alembic==1.14.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-settings==2.8.1
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.20
boto3==1.37.12
openai==1.68.2
pypdf==5.1.0
python-docx==1.1.2
pytest==8.3.5
httpx==0.28.1
